requirements.txt
langchain-community
duckduckgo-search
googlesearch-python
streamlit
langchain-google-genai
langchain-ollama
requests
